large changes to main file
implmented senors that would give priority to certain traffic light, giving the North-South light
the greatest prioirty if both are set.

seperated the control into two task where are both independent on a global state.

two mutex are used one for the shift register the other for the global states.